requirements.txt
pystray
pillow
win10toast
